feat: ensure compatibility for size_t to unsigned int conversions in Cap'n Proto serialization
- Updated `PlayerInput.hpp`, `GameState.hpp`, and `GameStart.hpp`:
  - Added `static_cast<unsigned int>` for size_t to resolve type mismatch warnings when initializing Cap'n Proto builders.
  - Ensured consistent and safe usage of indices in loop operations.
- Modified `tests/CMakeLists.txt`:
  - Adjusted DLL copying logic:
      - Replaced `copy_if_different` with a `true` command for placeholder functionality.
      - Updated to set runtime DLL paths in debugging environments using `VS_DEBUGGER_ENVIRONMENT`.

+++ common/Serialization/Capnp/Messages/C2S/PlayerInput.hpp
@@ -46,9 +46,9 @@ namespace RType::Messages::C2S {
 
             builder.setSequenceId(_sequenceId);
 
-            auto actionsBuilder = builder.initActions(actions.size());
+            auto actionsBuilder = builder.initActions(static_cast<unsigned int>(actions.size()));
             for (size_t i = 0; i < actions.size(); ++i) {
-                actionsBuilder.set(i, Shared::toCapnpAction(actions[i]));
+                actionsBuilder.set(static_cast<unsigned int>(i), Shared::toCapnpAction(actions[i]));
             }
 
             auto bytes = capnp::messageToFlatArray(message);
